server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
google-genai>=1.0.0
openai>=1.0.0
psycopg[binary]>=3.2.0
psycopg2-binary>=2.9.9
pandas>=2.2.0
numpy>=1.26.0
httpx>=0.27.0
tavily-python>=0.5.0
trafilatura>=2.0.0
lxml_html_clean>=0.4.1
boto3>=1.35.0
Pillow>=11.0.0
